Detect CPUs with throttled 512-bit width instructions
When compiling with `--instruction-set:native` detect the CPUs that implemented AVX-512 in a way where it's better not to use it much.

This replicates the check in codeman.cpp and passes the same flag to JIT that the VM would pass.

I noticed this because passing `--instruction-set:native` to our ASP.NET benchmark machines murders the performance (minus 10+% in RPS) instead of helping it in any way. Passing `--application.environmentVariables DOTNET_PreferredVectorBitWidth=512` kills it in similar way with CoreCLR-JIT, so this detection will likely fix it.