remove static linking support
it wasn't a good fit for CoreCLR because the EventPipe and GC symbols
clashed between CoreCLR and NativeAOT during static linking.

A possible way forward is to use incremental linking (`clang -Wl,-r`) on Linux/Mac in
combination with the `--exclude-libs` or `-hidden-lx` linker options.

However on Windows since NativeAOT static artifacts are always built
in Release mode, we would only be able to do static linking in limited
circumstances (and we would need to find the right linker option to do
the apropriate symbol hiding / incremental linking)