Derive OrderedEnumerable from Iterator (#98874)
* Move OrderedEnumerable to be nested in Enumerable

* Derive OrderedEnumerable from Iterator

The OrderedEnumerable implementation today doesn't derive from Iterator, and it has a GetEnumerator method implemented with yield. That means the optimization that allows the enumerable to be reused as the enumerator doesn't kick in, and we end up allocating a separate enumerator object in order to iterate through an Order{By} enumerable.