Fix the new EH with hot-cold split code
The new EH was not taking into account the fact that with hot-cold split
methods, the offset used for looking up EH funclets need to be computed
as if the hot and cold regions were consecutive in memory. That have
caused failures in a number of tests when the tests themselves were
compiled with R2R and hot-cold split enabled.

Close #98915, #98916, #98917, #98918