JIT: start checking sum of successor likelihoods
Add the ability to check that uccessor edge likelihoods sum to 1.0
Verify that successor edge likelihoods sum to 1.0 for the phases up through
and including `fgAddInternal`.

Fixed a few problems
* importer branch folding needs to update likelihoods
* gdv likelihood needed some adjustments, especially for the multi-guess case
* profile incorporation for OSR methods also needed some fixing. Here
we may have inconsistencies so the right answer is not always clear.
With this change we now trust the successor edge counts even if they
do not sum to the block count.

Contributes to #93020.