Disable the new exception handling by default again (#99066)
The new EH was enabled by default right before we've snapped for preview
2, but the various nightly CI tests that were executed after has
uncovered a couple of issues that I've never seen before. I've fixed two
of them, but another one has surfaced today, this time a GC hole that
seems to be related to the new EH. Since there is no time left for
inverstigating and fixing that for preview 2, I am disabling it by
default again.