Fix regression in jiterpreter trace generation
When I migrated the jiterpreter to a new branch target scanning pass I failed to filter the branch target tables it generates properly. This combined with an existing bug in the cfg code meant that traces in large methods ended up having big branch dispatch tables at the top where most of the branch targets were never hit, eating up a valuable chunk of the trace's 4KB limit.

This PR adds appropriate filtering in both the scanning pass and the cfg so that the dispatch table is as small as possible.