Fix unhandled exception when dependency is missing
ASPNet team has found that the new exception handling doesn't handle
unhandled exception caused by a missing dependency assembly correctly.
Instead of invoking the unhandled exception machinery in the native
code, it actually exited back to the managed exception handler loop with
a failure from the the SfiInit.

The reason is that in that specific case, the check if there are still
managed frames on the stack that's done by checking if there is any
explicit frame doesn't work, since there is a helper method frame.

This change fixes it by moving the unhandled exception check to the end
of the SfiInit when we haven't found any managed frame through the stack
walker.