SPMI: Avoid reading entire details CSV file into memory (#99116)
Switch to a streaming approach when processing the details CSV file to
save on the peak memory usage. We still need to save the rows with diffs
to be able to be able to pick examples to disassembly, but that list is
expected to be way smaller than the full set of methods replayed.