Add liveVarUpperSave
`liveVarUpperSave` will give a hint to the allocator that the given refposition
is something that was added conservatively and might not be live. Allocator when
sees this refposition and if the corresponding lclVarInterval was not assigned a
register, it will skip assiging or deallocating register for the current Refposition
as well. That way, the end effect of the extra RefPositions added for non-liveVars
is reduced to minimum.