SPMI: Project out fields for diffs information (#99226)
When there are many diffs we are still loading all of them into memory,
and in CI this can cause OOMs in these cases. Every row becomes a
dictionary that has an extra copy of all column names, and with the new
metrics work there are a lot of columns. Fix the problem by projecting
out just what is necessary from each row.