Fix stack walking for new EH
There is an edge case when the StackFrameIterator was incorrectly
setting the fShouldParentFrameUseUnwindTargetPCforGCReporting, which
resulted in a 100% reproducible GC hole in the
baseservices\exceptions\unittests\ThrowInFinallyNestedInTry with GC
stress C and tiered compilation off.
The fix is to delete code that is already present in an "if" branch
above the change and that should really be executed only when the
funclet parent is the caller of the actual handler frame.