Fix overaggressive switch optimization in regex source generator
We recently updated the source generator to be able to employ its alternation optimization that employs switch constructs in more places. However, it was being a bit overaggressive, in two ways:
1. Because of how it was slicing off the first character from a character loop, it was manufacturing a new node that the analysis engine hadn't seen before. As a result, it would report the node didn't contain any backtracking constructs, even if it did, and that would lead to braces possibly being emitted in the source where they shouldn't have been, which would lead to compilation failures when trying to goto a label inside of that block.
2. A concatenation of two nodes might lose the latter half of the concatenation if the first half isn't one of the few nodes we know how to handle.

Both of these come about from trying to avoid recomparing the single character at the beginning of the construct that was used in the switch. We can avoid the challenges by just doing that additional comparison, if necessary. In practice, the only case we can't do that where we were doing it is where the branch begins with a char loop.

(This was causing an outerloop test to fail, but it was masked by another recently-fixed issue that caused outerloop test failures to not be reported.)