JIT: Disallow parameters passed in float regs from being enregistered in int regs (#99286)
This change is a follow-up to 928ff3015d1936ca9985a0123754e13cbf47b237. The
logic in that change is only correct under the assumption that parameters passed
in float regs are never homed in integer regs. However, we can enregister
structs of fitting size in integer registers if they have only whole uses, so
that could still happen for a struct passed in float registers. This is a very
rare case because such enregistration only kicks in when the struct parameter
also was not promoted. This change disallows the problematic enregistration.