JIT: enable edge checks throughout
Fix the last remaining issues for edge likelihoods.

Main challenge here was switch lowering, particularly the expansions of switches
into a series of tests. The adjustments here are similar to those for multi-guess
GDV and type tests -- as we test possibilities one-by-one we have to adjust and
scale up likelihoods of remining possibilities. But for switches things are more
complex as edges may have dup counts, and we may eventually reach the point where
the remaining tests had zero initial likelihood.

Contributes to #93020.