Fix problem with scaling general loop blocks; add dumpers
1. Fix a problem where scaling block weights of a range of
basic blocks that compose a general loop could encounter a
block that is unreachable. This can happen for various kinds
of blocks that the JIT doesn't like to remove even if unreachable.
Simply skip scaling those blocks.

2. Add dumpers for `FlowGraphDfsTree` and `BlockToNaturalLoopMap`.
These can be called from the debugger (perhaps they should be called
to output these things to the JitDump).

3. Update `fgDfsBlocksAndRemove()` to dump any blocks that it did
not remove, even if they were unreachable.

This was found as part of fixing JitOptRepeat: https://github.com/dotnet/runtime/pull/94250