Remove bbFallsThrough checks in fgNewBBbefore/after (#99636)
Now that implicit fallthrough is gone, it doesn't make sense to mark a new block as rarely run if the previous block is rarely run, as there's no guarantee the previous block flows into the next. It would make more sense to rely on the weights of a block's predecessor edges to determine if it's rarely run.