Support frozen struct returns for Swift calls (#99704)
Adds support for pinvokes to Swift functions that return frozen structs in
multiple registers. This turned out to be simpler than I expected on the JIT
side; there is a small change necessary in `genMultiRegStoreToLocal` to take into
account that the Swift fields are going into offsets that don't necessarily
correspond to the register sizes (we already DNER the cases where things don't
work out, it seems).

Also adds 100 tests.

The support is complicated by the fact that Swift calls take the ret buffer in
rax on x64. This requires some VM side changes to avoid using rax in the interop
thunks.