JIT: Type fat pointer locals precisely, and avoid unnecessary normalization in inlining
During call importation, for fat pointer calls we will introduce a local
and spill the call to it. This loses track of the small typedness of the
value, which can cause inlining to introduce unnecessary normalization
casts later. For tailcalls this can cause us to add IR after the call
that we do not expect, causing issues like #99798.

Fix the problem by enhancing logic in a few places:
- Make the local created for these fat pointer calls small typed like
  regular normalize-on-store locals
- Enhance `fgCastNeeded` to take into account the small-typedness of
  these locals (like `IntegralRange::ForNode`)