Trivial performance suggestion for Linq `Where`
When the `Where` Linq extension method receives an empty list as the input, we can add a `Count == 0` check similar to arrays. If the incoming list is empty - why not just return this empty list and save allocating an extra `ListWhereIterator`.

I did not file an issue since this is a trivial change, sorry.

Feel free tor reject if this change is against the guidelines/vision, in that case - apologies.