Remove copy_version_files.sh
Arcade provides this via the GenerateNativeVersionFile target, which we
already call in various places. Lets remove this file.

This file, if used, will also cause issues in a VMR build without a git
repo. In that mode, git commands do not work, but Arcade and Source Link
will still provide git commit and related information correctly.