Process.Linux: handle when /proc and the process pid namespace don't match.
Normally the '/proc' filesystem uses the same pid namespace as the process.
With rootless containers, it may happen that these pid namespaces do not match
because the container doesn't have permissions to change '/proc' but it can
create a new pid namespace.

When that happens, the numeric ids used by the /proc filesystem no longer match with
the process pid namespace. We can still access information for the current process
using '/proc/self'. For other processes, we can't map pids to the proc pids, which
leads to reading information for non-existing/wrong/inaccessible processes.